New feature "Orchard.MediaLibrary.LocalizationExtensions": - MediaLocalizationMigrations: Adds LocalizationPart to all Media stereotyped types - MediaLibraryPickerFieldLocalizationExtensionHandler: Updates the media items within the field according to field settings - Settings for translation behaviours for MediaLibraryPickerField - Adds warning on localizing media items - form filter media by culture
"MediaLibrary" feature:
- Implemented Cloning methods of all medias PartDrivers
- MediaPartHandler: Adds a check to prevent file system media deletion when the file is referenced by other medias
- ClientStorageController: Adds a check to prevent file system media deletion when the file is referenced by other medias

"Localization" feature:
- Adds a logic to propagate returnurl parameter to the redirection url
- ILocalizationService: Adds a new method to set the culture of a content and at the same time the master content of it
- Propagate returnurl parameter if necessary

"Search" feature:
- updated json of media items called via ajax

# Conflicts:
#	src/Orchard.Web/Modules/Orchard.MediaLibrary/Controllers/ClientStorageController.cs
#	src/Orchard.Web/Modules/Orchard.MediaLibrary/Orchard.MediaLibrary.csproj

+++ src/Orchard.Web/Modules/Orchard.MediaLibrary/Controllers/ClientStorageController.cs
@@ -176,8 +176,18 @@ namespace Orchard.MediaLibrary.Controllers {
                     if (!replaceContentType.Equals(replaceMedia.TypeDefinition.Name, StringComparison.OrdinalIgnoreCase))
                         throw new Exception(T("Cannot replace {0} with {1}", replaceMedia.TypeDefinition.Name, replaceContentType).Text);
 
-                    _mediaLibraryService.DeleteFile(replaceMedia.FolderPath, replaceMedia.FileName);
-                    _mediaLibraryService.UploadMediaFile(replaceMedia.FolderPath, replaceMedia.FileName, file.InputStream);
+                    var mediaItemsUsingTheFile = Services.ContentManager.Query<MediaPart, MediaPartRecord>()
+                                                                .ForVersion(VersionOptions.Latest)
+                                                                .Where(x => x.FolderPath == replaceMedia.FolderPath && x.FileName == replaceMedia.FileName)
+                                                                .Count();
+                    if (mediaItemsUsingTheFile == 1) { // if the file is referenced only by the deleted media content, the file too can be removed.
+                        _mediaLibraryService.DeleteFile(replaceMedia.FolderPath, replaceMedia.FileName);
+                    }
+                    else {
+                        // it changes the media file name
+                        replaceMedia.FileName = filename;
+                    }
+                    _mediaLibraryService.UploadMediaFile(replaceMedia.FolderPath, filename, file.InputStream);
                     replaceMedia.MimeType = mimeType;
 
                     // Force a publish event which will update relevant Media properties
